Opage In app browser functions complete
Other information pages is completed with all the required functions all the links now open through an in-app browser all links are set and tested. Added a new dependency need to make use of it in the other page like food and travel where there are possible in app browser extensions might be used.