Travel page done - 2 sub travel pages added
Travel page is done and is linked to the main page and 2 sub pages are done in it one is travel in duluth and other is foreign and international travel. Both of them are completely liked and are set up need to add images and update the nav bar along with others.